Add hotfix for obstacle calculation
If the transmission line between sender and receiver goes exactly through the intersection of two walls,
this code adds the affected walls to the set of intersections.
Afterwards, the wall at which the beam leaves the obstacle will also be added to the set of intersections.
Hence, we have an odd amount of affected walls which will cause an assertion error below.
Therefore, the assertion caused an error and the simulation broke.